% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCA_OLD.R
\name{grabFeatureIndices}
\alias{grabFeatureIndices}
\title{Find feature indices of dataset i in appended dataset X = X1, X2, ..., Xn}
\usage{
grabFeatureIndices(i, numFeaturesPerDataset)
}
\arguments{
\item{i}{Number of the dataset}

\item{numFeaturesPerDataset}{Vector containing ncol per dataset}
}
\value{
result Feature indices of dataset i in appended dataset X.
}
\description{
Find feature indices of dataset i in appended dataset X = X1, X2, ..., Xn
}
