% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAR_MAIN.R
\name{initialisePESCA}
\alias{initialisePESCA}
\title{Initialises many aspects to run PESCA}
\usage{
initialisePESCA(
  dataSets,
  y,
  dataTypes,
  lambdas,
  lambdas_y,
  penalty = "L2",
  opts = opts,
  response = NULL,
  alpha = alpha,
  rstartseed = rstartseed
)
}
\arguments{
\item{dataSets}{A list of datasets.}

\item{dataTypes}{Data type per dataset, string of letters: "G" = gaussian, "P" = poisson, "B" = binary.}

\item{lambdas}{A list of fitted lambdas, one for each dataset.}

\item{penalty}{Penalty type for the PESCA function. Default is L2.}
}
\value{
A list of the following: A, B, mu, Theta, X, W, loss, rho, and some useful variables.
}
\description{
Initialises many aspects to run PESCA
}
