% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCA_OLD.R
\name{pESCA_CV}
\alias{pESCA_CV}
\title{pESCA model selection based on cross validation error}
\usage{
pESCA_CV(
  dataSets,
  dataTypes,
  lambdas_CV = NULL,
  penalty = "L2",
  fun_concave = "gdp",
  opts = list()
)
}
\arguments{
\item{lambdas_CV}{a vector cotains a sequence of values of lambda}
}
\value{
This function returns a list contains the results of a pESCA mdoel. \itemize{
\item cvErrors_mat: a matrix contains the CV errors for the full data set and each
single data set;
\item inits: a list contains the initilizations of all the constructed models;
\item outs: a list contains the outputs of all the constructed models;
}
}
\description{
This function implements a missing value based CV model selection
approach for the pESCA model on mutliple data sets with the same data type.
The details can be found in  \url{https://arxiv.org/abs/1902.06241}.
}
\examples{
\dontrun{
result_CV <- pESCA_CV(dataSets, dataTypes,
                            lambdas_CV, penalty='L2', fun_concave='gdp', opts=opts)
}

}
