% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCAR-package.R
\docType{package}
\name{PESCAR-package}
\alias{PESCAR}
\alias{PESCAR-package}
\title{PESCAR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fred White \email{f.t.g.white@uva.nl}

}
\keyword{internal}
