% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PESCA_OLD.R
\name{splitDataset}
\alias{splitDataset}
\title{GRvdP, FW;slightly tweaked.....    CV issue here

initialisePESCA = function(dataSets, dataTypes, lambdas, penalty="L2", opts=list(), response = NULL){
#' Initialises many aspects to run PESCA
#'
#' @param dataSets A list of datasets.
#' @param dataTypes Data type per dataset, string of letters: "G" = gaussian, "P" = poisson, "B" = binary.
#' @param lambdas A list of fitted lambdas, one for each dataset.
#' @param penalty Penalty type for the PESCA function. Default is L2.
#'
#' @returns A list of the following: A, B, mu, Theta, X, W, loss, rho, and some useful variables.

numDatasets = length(dataSets)
numSamplesPerDataset = unlist(lapply(dataSets, nrow))
numFeaturesPerDataset = unlist(lapply(dataSets, ncol))
totalNumFeatures = sum(numFeaturesPerDataset)
numComponents = opts$R
alphas = opts$alphas
rand_start = opts$rand_start
rhos = rep(NA, numDatasets) # form rhos, the Lipschitz constant for each data typeform full data set X, X = X1,...Xl,...XLform full weighting matrix W, W = W1,...Wl,...WLbrowser()

X = data.matrix(do.call(cbind.data.frame, dataSets))
W = 1 * (!is.na(X))

W\link{is.na(X)} <- 0
X\link{is.na(X)} <- 0initialization

if(exists("A0", where=opts)){
mu = t(opts$mu0)
A = opts$A0
B = opts$B0
}
else if(rand_start == 1){ # use random initialization
mu = matrix(data=0, nrow=1, ncol=totalNumFeatures)
set.seed(1)
A = matrix(rnorm(numSamplesPerDataset\link{1}\emph{numComponents), nrow=numSamplesPerDataset\link{1}, ncol=numComponents)
B = matrix(rnorm(totalNumFeatures}numComponents), nrow=totalNumFeatures, ncol=numComponents)
}
else if(rand_start == 0 & !("P" \%in\% dataTypes)){ # use SCA model as initialization, Poisson distribution is not used
mu = matrix(colMeans(X), 1, totalNumFeatures)

\if{html}{\out{<div class="sourceCode">}}\preformatted{# browser()      ###fails with CV possibly due to NAs -> also need to not split block if only 1 feature... how to handle this?
X_svd = RSpectra::svds(scale(X,center=TRUE,scale=FALSE), numComponents, nu=numComponents, nv=numComponents)
A = X_svd$u
B = X_svd$v \%*\% diag(X_svd$d[1:numComponents])
}\if{html}{\out{</div>}}

}
else if(rand_start == 0){ # Poisson distribution version
X_tmp = X
for(i in 1:numDatasets){ #log transformation applied to Poisson data
dataType = dataTypes\link{i}
if(dataType == 'P'){
columns_Xi = grabFeatureIndices(i,numFeaturesPerDataset)
X_tmp\link{,columns_Xi} = log(X\link{,columns_Xi} + 1)
}
}
mu = matrix(colMeans(X_tmp), 1, totalNumFeatures)
X_svd = RSpectra::svds(scale(X_tmp,center=TRUE,scale=FALSE), numComponents, nu=numComponents, nv=numComponents)
A = X_svd$u
B = X_svd$v \%*\% diag(X_svd$d\link{1:numComponents})   ##why put size on the loadings not scores?
}

#FAILS CV not first iteration - FIXED - (?) - CHECK -
if(is.null(dim(mu)) | ncol(mu) == 1){
mu <- t(mu)
}browser()

Theta = RpESCA::ones(numSamplesPerDataset\link{1}) \%*\% mu + tcrossprod(A,B)Split up result to individual datasets again

Xs = splitDataset(X, numFeaturesPerDataset)
Ws = splitDataset(W, numFeaturesPerDataset)
Thetas = splitDataset(Theta, numFeaturesPerDataset)
Bs = splitDataset(as.matrix(B), numFeaturesPerDataset, direction="rows")Prepare rhos

for(i in 1:numDatasets){
dataType = dataTypes\link{i}
if(dataType == 'G') rhos\link{i} = 1
if(dataType == 'B') rhos\link{i} = 0.25
if(dataType == 'P') rhos\link{i} = max(exp(Thetas[\link{i}]))
}Calculate loss of initial estimation

if(is.null(response)){
# lossResult = calculatePESCAloss(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty)
lossResult = calculatePESCAloss(dataTypes = dataTypes,
Xs = Xs,
Ws = Ws,
Thetas = Thetas,
alphas = alphas,
lambdas = lambdas,
Bs = Bs,
numComponents = numComponents,
penalty = penalty)

}else{
lossResult = calculatePESCARloss(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty,  fun_concave = "gdp", response, A_new = A, Wt = Wt)   #CHANGE hardcoded
# calculatePESCARloss = function(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty="L2", fun_concave="gdp", response, A){

}

return(list("A" = A,                        # Scores of the SCA model
"B" = B,                        # Loadings of the SCA model
"mu" = mu,                      # Means of the SCA model
"Theta" = Theta,                # Reconstructed data of the SCA model
"X" = X,                        # Input datasets following X=\link{X1,...,XL}
"Xs" = Xs,                      # Input datasets in list format
"W" = W,                        # Weight matrix of X with element=1 is the value is non-zero
"Ws" = Ws,                      # Weight matrix of each dataset in list format
"Thetas" = Thetas,              # Reconstructed datasets of the SCA model in list format
"Bs" = Bs,                      # Loadings of the SCA model for each dataset, in list format
"loss" = lossResult,            # Loss of iteration 0
"rhos" = rhos,
"numDatasets" = numDatasets,
"numSamplesPerDataset" = numSamplesPerDataset,
"numFeaturesPerDataset" = numFeaturesPerDataset,
"totalNumFeatures" = totalNumFeatures))
}}
\usage{
splitDataset(df, numFeaturesPerDataset, direction = "columns")
}
\description{
GRvdP, FW;slightly tweaked.....    CV issue here

initialisePESCA = function(dataSets, dataTypes, lambdas, penalty="L2", opts=list(), response = NULL){
#' Initialises many aspects to run PESCA
#'
#' @param dataSets A list of datasets.
#' @param dataTypes Data type per dataset, string of letters: "G" = gaussian, "P" = poisson, "B" = binary.
#' @param lambdas A list of fitted lambdas, one for each dataset.
#' @param penalty Penalty type for the PESCA function. Default is L2.
#'
#' @returns A list of the following: A, B, mu, Theta, X, W, loss, rho, and some useful variables.

numDatasets = length(dataSets)
numSamplesPerDataset = unlist(lapply(dataSets, nrow))
numFeaturesPerDataset = unlist(lapply(dataSets, ncol))
totalNumFeatures = sum(numFeaturesPerDataset)
numComponents = opts$R
alphas = opts$alphas
rand_start = opts$rand_start
rhos = rep(NA, numDatasets) # form rhos, the Lipschitz constant for each data typeform full data set X, X = X1,...Xl,...XLform full weighting matrix W, W = W1,...Wl,...WLbrowser()

X = data.matrix(do.call(cbind.data.frame, dataSets))
W = 1 * (!is.na(X))

W\link{is.na(X)} <- 0
X\link{is.na(X)} <- 0initialization

if(exists("A0", where=opts)){
mu = t(opts$mu0)
A = opts$A0
B = opts$B0
}
else if(rand_start == 1){ # use random initialization
mu = matrix(data=0, nrow=1, ncol=totalNumFeatures)
set.seed(1)
A = matrix(rnorm(numSamplesPerDataset\link{1}\emph{numComponents), nrow=numSamplesPerDataset\link{1}, ncol=numComponents)
B = matrix(rnorm(totalNumFeatures}numComponents), nrow=totalNumFeatures, ncol=numComponents)
}
else if(rand_start == 0 & !("P" \%in\% dataTypes)){ # use SCA model as initialization, Poisson distribution is not used
mu = matrix(colMeans(X), 1, totalNumFeatures)

\if{html}{\out{<div class="sourceCode">}}\preformatted{# browser()      ###fails with CV possibly due to NAs -> also need to not split block if only 1 feature... how to handle this?
X_svd = RSpectra::svds(scale(X,center=TRUE,scale=FALSE), numComponents, nu=numComponents, nv=numComponents)
A = X_svd$u
B = X_svd$v \%*\% diag(X_svd$d[1:numComponents])
}\if{html}{\out{</div>}}

}
else if(rand_start == 0){ # Poisson distribution version
X_tmp = X
for(i in 1:numDatasets){ #log transformation applied to Poisson data
dataType = dataTypes\link{i}
if(dataType == 'P'){
columns_Xi = grabFeatureIndices(i,numFeaturesPerDataset)
X_tmp\link{,columns_Xi} = log(X\link{,columns_Xi} + 1)
}
}
mu = matrix(colMeans(X_tmp), 1, totalNumFeatures)
X_svd = RSpectra::svds(scale(X_tmp,center=TRUE,scale=FALSE), numComponents, nu=numComponents, nv=numComponents)
A = X_svd$u
B = X_svd$v \%*\% diag(X_svd$d\link{1:numComponents})   ##why put size on the loadings not scores?
}

#FAILS CV not first iteration - FIXED - (?) - CHECK -
if(is.null(dim(mu)) | ncol(mu) == 1){
mu <- t(mu)
}browser()

Theta = RpESCA::ones(numSamplesPerDataset\link{1}) \%*\% mu + tcrossprod(A,B)Split up result to individual datasets again

Xs = splitDataset(X, numFeaturesPerDataset)
Ws = splitDataset(W, numFeaturesPerDataset)
Thetas = splitDataset(Theta, numFeaturesPerDataset)
Bs = splitDataset(as.matrix(B), numFeaturesPerDataset, direction="rows")Prepare rhos

for(i in 1:numDatasets){
dataType = dataTypes\link{i}
if(dataType == 'G') rhos\link{i} = 1
if(dataType == 'B') rhos\link{i} = 0.25
if(dataType == 'P') rhos\link{i} = max(exp(Thetas[\link{i}]))
}Calculate loss of initial estimation

if(is.null(response)){
# lossResult = calculatePESCAloss(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty)
lossResult = calculatePESCAloss(dataTypes = dataTypes,
Xs = Xs,
Ws = Ws,
Thetas = Thetas,
alphas = alphas,
lambdas = lambdas,
Bs = Bs,
numComponents = numComponents,
penalty = penalty)

}else{
lossResult = calculatePESCARloss(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty,  fun_concave = "gdp", response, A_new = A, Wt = Wt)   #CHANGE hardcoded
# calculatePESCARloss = function(dataTypes, Xs, Ws, Thetas, alphas, lambdas, Bs, numComponents, penalty="L2", fun_concave="gdp", response, A){

}

return(list("A" = A,                        # Scores of the SCA model
"B" = B,                        # Loadings of the SCA model
"mu" = mu,                      # Means of the SCA model
"Theta" = Theta,                # Reconstructed data of the SCA model
"X" = X,                        # Input datasets following X=\link{X1,...,XL}
"Xs" = Xs,                      # Input datasets in list format
"W" = W,                        # Weight matrix of X with element=1 is the value is non-zero
"Ws" = Ws,                      # Weight matrix of each dataset in list format
"Thetas" = Thetas,              # Reconstructed datasets of the SCA model in list format
"Bs" = Bs,                      # Loadings of the SCA model for each dataset, in list format
"loss" = lossResult,            # Loss of iteration 0
"rhos" = rhos,
"numDatasets" = numDatasets,
"numSamplesPerDataset" = numSamplesPerDataset,
"numFeaturesPerDataset" = numFeaturesPerDataset,
"totalNumFeatures" = totalNumFeatures))
}
}
